grpc_service/requirements.txt
grpcio>=1.71.0
grpcio-tools
protobuf
asyncpg
SQLAlchemy>=2.0
alembic
python-dotenv
